healthserver/nutrition-service/requirements.txt

# DEPENDANCES FLASK & UTILITAIRES

Flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
pymongo==4.15.3
python-dotenv==1.2.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6
click==8.1.7
rich==14.2.0
blinker>=1.9.0
markdown-it-py>=2.2.0
pygments>=2.13.0,<3.0.0

# MINIO & CRYPTOGRAPHIE

minio==7.2.18
pycryptodome==3.23.0
argon2-cffi

# MACHINE LEARNING & IA

tensorflow==2.16.1
numpy==1.24.3
opencv-python==4.9.0.80
h5py==3.15.1
pillow==12.0.0
keras==3.12.0

# DEPENDANCES TECHNIQUES TENSORFLOW

absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0

opt_einsum==3.4.0
optree==0.17.0
tensorboard==2.16.1
tensorboard-data-server==0.7.2
termcolor==3.2.0
typing_extensions==4.15.0
wrapt==2.0.0
protobuf==3.20.3

# AUTRES LIBRAIRIES UTILES
packaging==25.0
idna==3.11
charset-normalizer==3.4.4
urllib3==2.5.0
certifi==2025.10.5
six==1.17.0
dnspython==2.8.0
